move primitive files to _legacy
