+++ .gitignore
@@ -1,5 +1,1 @@
-node_modules
 .DS_Store
-dist
-dist-ssr
-*.local
