--- README.md
@@ -1,5 +1,0 @@
-[![Netlify Status](https://api.netlify.com/api/v1/badges/b73c4b3d-1f8e-4827-8027-a8b3c7fe1a58/deploy-status)](https://app.netlify.com/sites/owencode/deploys)
-
-# OwenCode
-
-Description coming soon.
