+++ _legacy/.prettierignore
@@ -1,0 +1,2 @@
+build
+coverage
