+++ _legacy/postcss.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  plugins: {
+    "postcss-preset-env": { stage: 1 },
+    tailwindcss: {},
+    autoprefixer: {},
+  },
+};
