+++ _legacy/vite.config.js
@@ -1,0 +1,7 @@
+import { defineConfig } from "vite";
+
+export default defineConfig({
+  server: {
+    host: true,
+  },
+});
