--- index.html
@@ -1,161 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="favicon.svg" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>OwenCode</title>
-  </head>
-
-  <body>
-    <canvas id="bg"></canvas>
-
-    <main id="content">
-      <section class="justify-end">
-        <div>
-          <h1 class="font-mono font-bold text-7xl">OwenCode</h1>
-        </div>
-      </section>
-  
-      <section>
-        <div>
-          <h2>Who Am I?</h2>
-          <p>
-            My name is Owen Bickford. I am a full stack web developer based in Michigan and working
-            remotely.
-          </p>
-  
-          <p>
-            This static page is where I keep track of tech I have used, stuff I have worked on, and where I experiment with fun stuff like Three JS.
-          </p>
-        </div>
-      </section>
-  
-      <section>
-        <div>
-          <h2>Define Full Stack</h2>
-          <p>
-            My full stack <em>experience</em> includes the following technologies:
-  
-            <ul class="flex flex-wrap gap-6">
-              <li>
-                <a href="https://www.mysql.com/">MySQL</a>
-              </li>
-
-              <li>
-                <a href="https://www.postgresql.org/">Postgres</a>
-              </li>
-
-              <li>
-                <a href="https://www.oracle.com/database/">Oracle</a>
-              </li>
-
-              <li>
-                <a href="https://www.oracle.com/database/">Presto SQL</a>
-              </li>
-
-              <li>
-                <a href="https://redis.io/">Redis</a>
-              </li>
-
-              <li>
-                <a href="https://www.php.net/">PHP</a>
-              </li>
-
-              <li>
-                <a href="https://laravel.com/">Laravel</a>
-              </li>
-
-              <li>
-                <a href="https://laravel-livewire.com/">Laravel Livewire</a>
-              </li>
-
-              <li>
-                <a href="https://developer.mozilla.org/en-US/docs/Web/JavaScript">Javascript</a>
-              </li>
-
-              <li>
-                <a href="https://github.com/alpinejs/alpine">AlpineJS</a>
-              </li>
-
-              <li>
-                <a href="https://vuejs.org/">VueJS</a>
-              </li>
-
-              <li>
-                <a href="https://tailwindcss.com/">TailwindCSS</a>
-              </li>
-
-              <li>
-                <a href="https://getbootstrap.com/">Bootstrap</a>
-              </li>
-
-              <li>
-                <a href="https://elixir-lang.org/">Elixir</a>
-              </li>
-
-              <li>
-                <a href="https://www.phoenixframework.org/">Phoenix</a>
-              </li>
-
-              <li>
-                <a href="https://hexdocs.pm/phoenix_live_view/Phoenix.LiveView.html">Phoenix LiveView</a>
-              </li>
-
-              <li>
-                <a href="https://www.cloudfoundry.org/">Cloud Foundry</a>
-              </li>
-
-              <li>
-                <a href="https://www.redhat.com/en/topics/api/what-is-a-rest-api">REST</a>
-              </li>
-
-              <li>
-                <a href="https://graphql.org/">GraphQL</a>
-              </li>
-
-              <li>
-                <a href="https://oauth.net/2/">OAuth</a>
-              </li>
-
-              <li>
-                <a href="https://openid.net/connect/">OpenID Connect</a>
-              </li>
-
-              <li>
-                <a href="https://docs.microsoft.com/en-us/azure/active-directory/hybrid/how-to-connect-sso#:~:text=Azure%20Active%20Directory%20Seamless%20Single%20Sign%2DOn%20(Azure%20AD%20Seamless,even%20type%20in%20their%20usernames.">Azure SSO</a>
-              </li>
-
-              <li>
-                <a href="https://getfedora.org/">Linux</a>
-              </li>
-
-            </ul>
-          </p>
-        </div>
-      </section>
-  
-      <section>
-        <div>
-          <h2>What Do I Enjoy?</h2>
-          <p>
-            I enjoy solving problems and optimizing those solutions.
-          </p>
-        </div>
-      </section>
-      
-      <section>
-        <div>
-          <h2>Reach Out</h2>
-          <ul class="flex flex-wrap gap-4">
-            <li><a href="https://github.com/type1fool/">GitHub</a></li>
-            <li><a href="https://twitter.com/owenbickford">Twitter</a></li>
-          </ul>
-        </div>
-      </section>
-    </main>
-  
-
-    <script type="module" src="/main.js"></script>
-  </body>
-</html>
