--- package.json
@@ -1,22 +1,0 @@
-{
-  "name": "owencode",
-  "version": "0.0.0",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "serve": "vite preview"
-  },
-  "devDependencies": {
-    "@tailwindcss/forms": "^0.3.2",
-    "@tailwindcss/typography": "^0.4.0",
-    "autoprefixer": "^10.2.5",
-    "postcss": "^8.3.0",
-    "postcss-preset-env": "^6.7.0",
-    "prettier": "2.3.0",
-    "tailwindcss": "^2.1.2",
-    "vite": "^2.3.3"
-  },
-  "dependencies": {
-    "three": "^0.128.0"
-  }
-}
