--- postcss.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  plugins: {
-    "postcss-preset-env": { stage: 1 },
-    tailwindcss: {},
-    autoprefixer: {},
-  },
-};
