--- vite.config.js
@@ -1,7 +1,0 @@
-import { defineConfig } from "vite";
-
-export default defineConfig({
-  server: {
-    host: true,
-  },
-});
